Metrics: Removing separate namespace.
