+++ Test/CoreSDK.Test/Shared/Core.Shared.Tests.projitems
@@ -32,12 +32,11 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\TelemetryProcessorChainBuilderTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\Tracing\CoreEventSourceTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\Tracing\ExtensionsTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Metrics\RequestMetricExtractorTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Metrics\MetricExtractorTelemetryProcessorBaseTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Metrics\MetricManagerTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Metrics\MetricSample.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Metrics\MetricTest.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Metrics\StubMetricProcessor.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\MetricExtractorTelemetryProcessorBaseTest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\MetricManagerTest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\MetricSample.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\MetricTest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\RequestMetricExtractorTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\SequencePropertyInitializerTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\ComponentContextTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\DeviceContextTest.cs" />
@@ -72,6 +71,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\Tracing\ThreadResourceLockTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\UserContextTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\Implementation\WeakConcurrentRandomTests.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Extensibility\StubMetricProcessor.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\TelemetryConfigurationFactoryTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensibility\TelemetryConfigurationTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)OperationTelemetryExtensionsTests.cs" />
