+++ Test/CoreSDK.Test/Shared/Extensibility/MetricSample.cs
