+++ Test/CoreSDK.Test/Shared/Extensibility/MetricTest.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Microsoft.ApplicationInsights.Extensibility.Metrics
+﻿namespace Microsoft.ApplicationInsights.Extensibility
 {
     using System;
     using System.Collections.Generic;
