+++ src/Core/Managed/Shared/Extensibility/Implementation/SimpleMetricStatisticsAggregator.cs
