+++ src/TelemetryChannels/ServerTelemetryChannel/NuGet/ApplicationInsights.config.install.xdt
@@ -10,10 +10,10 @@
   </TelemetryProcessors>
 
   <TelemetryProcessors xdt:Transform="InsertIfMissing">
-    <Add Type="Microsoft.ApplicationInsights.Extensibility.Metrics.RequestMetricExtractor, Microsoft.ApplicationInsights"
+    <Add Type="Microsoft.ApplicationInsights.Extensibility.RequestMetricExtractor, Microsoft.ApplicationInsights"
          xdt:Transform="InsertIfMissing"
          xdt:Locator="Match(Type)"></Add>
-    <Add Type="Microsoft.ApplicationInsights.Extensibility.Metrics.DependencyMetricExtractor, Microsoft.ApplicationInsights"
+    <Add Type="Microsoft.ApplicationInsights.Extensibility.DependencyMetricExtractor, Microsoft.ApplicationInsights"
         xdt:Transform="InsertIfMissing"
         xdt:Locator="Match(Type)"></Add>
     <Add Type="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel.AdaptiveSamplingTelemetryProcessor, Microsoft.AI.ServerTelemetryChannel"
